requirements.txt
ipython
jupyter
chardet
pyyaml
pylint
pytest
pytest-cov
